Move tiled getRegion tests to GDAL only
